# lc-ssi-wc/openapi/swift-data-service.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baseline SWIFT Data Service",
    "version": "1.1.0-prototype",
    "description": "SR2026 prototype contract. BIC syntax follows ISO 9362; currency codes are supplied by the ISO 4217 Currency service. Licensed SwiftRef/RMA/SSI data remains a production prerequisite."
  },
  "servers": [
    {
      "url": "http://localhost:3100/api"
    }
  ],
...
  "paths": {
    "/rma-authorisations": {
      "get": {
        "operationId": "listRma",
        "responses": {
          "200": {
            "description": "RMA records"
          }
        }
      },
      "post": {
        "operationId": "createRmaDraft",
        "requestBody": {
          "$ref": "#/components/requestBodies/Rma"
        },
        "responses": {
          "201": {
            "description": "DRAFT created"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/rma-authorisations/{id}": {
      "put": {
        "operationId": "updateRmaDraft",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Rma"
        },
        "responses": {
          "200": {
            "description": "DRAFT updated"
          }
        }
      },
      "delete": {
        "operationId": "revokeRma",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Logically revoked; audit retained"
          }
        }
      }
    },
...
    "/ssis": {
      "get": {
        "operationId": "listSsi",
        "responses": {
          "200": {
            "description": "SSI records"
          }
        }
      },
      "post": {
        "operationId": "createSsiDraft",
        "requestBody": {
          "$ref": "#/components/requestBodies/Ssi"
        },
        "responses": {
          "201": {
            "description": "DRAFT created"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/ssis/{id}": {
      "put": {
        "operationId": "updateSsiDraft",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Ssi"
        },
        "responses": {
          "200": {
            "description": "DRAFT updated"
          }
        }
      },
      "delete": {
        "operationId": "revokeSsi",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Logically revoked; audit retained"
          }
        }
      }
    },
    "/nostro-accounts": {
      "get": {
        "operationId": "listNostro",
        "responses": {
          "200": {
            "description": "Nostro records"
          }
        }
      },
      "post": {
        "operationId": "createNostroDraft",
        "requestBody": {
          "$ref": "#/components/requestBodies/Nostro"
        },
        "responses": {
          "201": {
            "description": "DRAFT created"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/nostro-accounts/{id}": {
      "put": {
        "operationId": "updateNostroDraft",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Nostro"
        },
        "responses": {
          "200": {
            "description": "DRAFT updated"
          }
        }
      },
      "delete": {
        "operationId": "revokeNostro",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Logically revoked; audit retained"
          }
        }
      }
    },
...
    "/reference/fin-tag-suggestions": {
      "post": {
        "operationId": "suggestReferenceFin5xTags",
        "description": "Deprecated compatibility adapter. Use POST /reference/fin-tag-resolutions. It maps resolvedFields to the legacy suggestions shape and never supports payment execution.",
        "responses": {
          "200": {
            "description": "Non-payment MT 5x suggestions; paymentExecutable is always false."
          },
          "400": {
            "description": "FIN context incomplete or no verified exact mapping."
          }
        },
        "deprecated": true
      }
    },
    "/settlements/resolve": {
      "post": {
        "operationId": "previewExecutableSettlement",
        "description": "Pre-settlement PAYMENT_SSI preview for an OUTGOING executable MT2 payment. The server derives counterpartyType=BANK from the selectable SR2026 Payment Message Index profile and rejects a client-supplied counterpartyType. MT/MX output format is deliberately absent because it is not a matching or ranking key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mt2BankResolutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked preview; paymentExecutable remains false until confirm.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mt2ResolutionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Incomplete or non-payment/incoming context, unsupported Message Index profile, target mismatch, or client-supplied counterpartyType."
          }
        }
      }
    },
    "/settlements/{resolutionId}/confirm": {
      "post": {
        "operationId": "confirmExecutableSettlement",
        "parameters": [
          {
            "name": "resolutionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Rechecks eligibility, Actual Receiver RMA and matching Nostro, then returns canonical roles plus immutable snapshot/token/hash. The caller owns full message assembly, transaction lifecycle and accounting.",
        "responses": {
          "200": {
            "description": "Confirmed PAYMENT_SSI; paymentExecutable true."
          },
          "409": {
            "description": "Preview stale or route no longer eligible."
          }
        }
      }
    },
    "/settlements/clearing-options": {
      "post": {
        "operationId": "listEligibleResolutionClearingOptions",
        "description": "Returns only clearing systems referenced by currently Active and eligible SSI routes for the submitted MT2 payment-trigger transaction criteria, intersected with active Clearing System Standing Data. The selectable sourceMessageType is required; counterpartyType=BANK is server-derived and must not be supplied by the client. This is not the full capability catalogue.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mt2BankResolutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Eligible SSI-backed clearing override options.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mt2ResolutionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/own-account-settlements/resolve": {
      "post": {
        "operationId": "resolveOwnAccountSettlement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnAccountSettlementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MRG-governed own-account result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mt2ResolutionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/fi-direct-debits/resolve": {
      "post": {
        "operationId": "resolveFiDirectDebit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FiDirectDebitRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MRG-governed FI direct-debit result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mt2ResolutionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/payment-notifications/resolve": {
      "post": {
        "operationId": "resolvePaymentNotification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentNotificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MRG-governed payment-notification result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mt2ResolutionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reference/clearing-systems": {
      "get": {
        "operationId": "listClearingSystems",
        "parameters": [
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Currency"
            }
          },
          {
            "name": "settlementCountry",
            "in": "query",
            "required": false,
            "description": "ISO 3166-1 alpha-2 market hint. Includes exact-market rails and governed PAN_REGIONAL rails whose eligibleCountries contain the requested country.",
            "schema": {
              "type": "string",
              "pattern": "^[A-Z]{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active governed clearing-system standing data; TARGET/TARGET2 are legacy aliases for T2.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearingSystemPage"
                }
              }
            }
          }
        }
      }
    },
    "/reference/fin-resolution-catalogue": {
      "get": {
        "operationId": "listFinResolutionCatalogue",
        "description": "Returns the SR-governed outward MT3/MT4/MT7 catalogue with an explicit message-level SSI resolution scope marker and SSI-resolvable tag options.",
        "parameters": [
          {
            "name": "standardsRelease",
            "in": "query",
            "schema": { "type": "string", "default": "SR2026" }
          }
        ],
        "responses": {
          "200": {
            "description": "Parameter-driven FIN Resolution catalogue; each message is marked SSI_RESOLVABLE or NO_SSI_RESOLVABLE_FIELDS."
          }
        }
      }
    },
...
    "/v1/resolution-page-definitions/index": {
      "get": {
        "summary": "List governed resolution page definitions",
        "parameters": [
          {
            "name": "standardsRelease",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "businessDomain",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageParameterBusinessDomain"
            }
          }
        ],
        "responses": {
          "200": { "description": "Governed definition index" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "requestBodies": {
      "Rma": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RmaCommand"
            }
          }
        }
      },
      "Ssi": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SsiCommand"
            }
          }
        }
      },
      "Nostro": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NostroCommand"
            }
          }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "SWIFT/ISO syntax or business-rule validation failed"
      }
    },
    "schemas": {
...
      "Mt2ResolutionResponse": {
        "type": "object",
        "required": ["redirectDomain"],
        "properties": {
          "redirectDomain": {
            "type": ["string", "null"],
            "description": "Owning MT2 resolution domain when the caller must redirect; null on a terminal decision."
          }
        }
      },
      "Mt2BankResolutionRequest": {
        "type": "object",
        "required": ["counterpartyBankServiceId"],
        "properties": {
          "counterpartyBankServiceId": {
            "$ref": "#/components/schemas/BankServiceId"
          },
          "sourceMessageType": { "type": "string" },
          "messageType": { "type": "string" }
        },
        "description": "The service resolves the read-only BIC from the Bank Service identity; callers must not submit a BIC."
      },
...
    }
  }
}